+++ testutil/sims/simulation_helpers.go
@@ -1,18 +1,21 @@
 package sims
 
 import (
+	"bytes"
 	"encoding/json"
 	"fmt"
 	"os"
 
 	dbm "github.com/cosmos/cosmos-db"
+	"github.com/tendermint/tendermint/libs/log"
+
 	"github.com/cosmos/cosmos-sdk/codec"
 	"github.com/cosmos/cosmos-sdk/runtime"
+	storetypes "github.com/cosmos/cosmos-sdk/store/types"
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	"github.com/cosmos/cosmos-sdk/types/kv"
 	"github.com/cosmos/cosmos-sdk/types/module"
 	simtypes "github.com/cosmos/cosmos-sdk/types/simulation"
-	"github.com/tendermint/tendermint/libs/log"
 )
 
 // SetupSimulation creates the config, db (levelDB), temporary directory and logger for the simulation tests.
@@ -106,7 +109,7 @@ func PrintStats(db dbm.DB) {
 
 // GetSimulationLog unmarshals the KVPair's Value to the corresponding type based on the
 // each's module store key and the prefix bytes of the KVPair's key.
-func GetSimulationLog(storeName string, sdr sdk.StoreDecoderRegistry, kvAs, kvBs []kv.Pair) (log string) {
+func GetSimulationLog(storeName string, sdr storetypes.StoreDecoderRegistry, kvAs, kvBs []kv.Pair) (log string) {
 	for i := 0; i < len(kvAs); i++ {
 		if len(kvAs[i].Value) == 0 && len(kvBs[i].Value) == 0 {
 			// skip if the value doesn't have any bytes
@@ -122,4 +125,60 @@ func GetSimulationLog(storeName string, sdr sdk.StoreDecoderRegistry, kvAs, kvBs
 	}
 
 	return log
+}
+
+// DiffKVStores compares two KVstores and returns all the key/value pairs
+// that differ from one another. It also skips value comparison for a set of provided prefixes.
+func DiffKVStores(a storetypes.KVStore, b storetypes.KVStore, prefixesToSkip [][]byte) (kvAs, kvBs []kv.Pair) {
+	iterA := a.Iterator(nil, nil)
+
+	defer iterA.Close()
+
+	iterB := b.Iterator(nil, nil)
+
+	defer iterB.Close()
+
+	for {
+		if !iterA.Valid() && !iterB.Valid() {
+			return kvAs, kvBs
+		}
+
+		var kvA, kvB kv.Pair
+		if iterA.Valid() {
+			kvA = kv.Pair{Key: iterA.Key(), Value: iterA.Value()}
+
+			iterA.Next()
+		}
+
+		if iterB.Valid() {
+			kvB = kv.Pair{Key: iterB.Key(), Value: iterB.Value()}
+		}
+
+		compareValue := true
+
+		for _, prefix := range prefixesToSkip {
+			// Skip value comparison if we matched a prefix
+			if bytes.HasPrefix(kvA.Key, prefix) {
+				compareValue = false
+				break
+			}
+		}
+
+		if !compareValue {
+			// We're skipping this key due to an exclusion prefix.  If it's present in B, iterate past it.  If it's
+			// absent don't iterate.
+			if bytes.Equal(kvA.Key, kvB.Key) {
+				iterB.Next()
+			}
+			continue
+		}
+
+		// always iterate B when comparing
+		iterB.Next()
+
+		if !bytes.Equal(kvA.Key, kvB.Key) || !bytes.Equal(kvA.Value, kvB.Value) {
+			kvAs = append(kvAs, kvA)
+			kvBs = append(kvBs, kvB)
+		}
+	}
 }