+++ x/upgrade/keeper/migrations_test.go
@@ -4,8 +4,8 @@ import (
 	"encoding/binary"
 	"testing"
 
+	storetypes "github.com/cosmos/cosmos-sdk/store/types"
 	"github.com/cosmos/cosmos-sdk/testutil"
-	sdk "github.com/cosmos/cosmos-sdk/types"
 	"github.com/cosmos/cosmos-sdk/x/upgrade/types"
 	"github.com/stretchr/testify/require"
 )
@@ -20,8 +20,8 @@ func encodeOldDoneKey(upgrade storedUpgrade) []byte {
 }
 
 func TestMigrateDoneUpgradeKeys(t *testing.T) {
-	upgradeKey := sdk.NewKVStoreKey("upgrade")
-	ctx := testutil.DefaultContext(upgradeKey, sdk.NewTransientStoreKey("transient_test"))
+	upgradeKey := storetypes.NewKVStoreKey("upgrade")
+	ctx := testutil.DefaultContext(upgradeKey, storetypes.NewTransientStoreKey("transient_test"))
 	store := ctx.KVStore(upgradeKey)
 
 	testCases := []struct {